Update parser for new format
The format of the dumps produced by xEdit has changed since I first
wrote this code.

Instead of updating the parser code to account for the changes, we've
instead switched to a custom format that looks like this:

<INFO stuff> in <DIAL stuff> in <QUST stuff>

where

INFO stuff is an optional editor ID in quotes followed by
"[INFO <Form ID>]" where <Form ID> is replaced with the 8-character form
ID.

DIAL stuff and QUST stuff follow the same pattern except for of course
the characters in the square brackets will be DIAL and QUST.

After this line will be lines beginning with the response number (NOT
the response index) and then the text of each of the reponses in the
INFO.

This greatly simplifies the parsing code and should make things less
likely to break in the future.

Also in this commit, I added more context info to the prompt that
checks to see if two responses are the same across the two files. There
may be more we want to do here to help resolve the question, but I was
finding it hard to answer the question in some cases because the
responses did look similar but not entirely close, either.