Improve parser and fix bugs
The file parser code now uses more LINQ to simplify the code further.

This commit fixes a bug in parsing the quest editor ID for an INFO. I
had added something to the regex previously to grab the "in" text before
the editor ID for the quest. I had made a similar change to the DIAL
parsing, but when I found a better way to take a smaller match I removed
it from there. I just missed removing it from the QUST part.

I also added a prompt so that you don't absolutely have to try to
resolve the "maybe equal" responses every time you load two files. When
this tool is evolved to be able to produce an output file that the
Skyblivion installer will use as input, we will want to make sure to
resolve all of the maybes, but for now it can be a nuisance if you're
not specifically testing that part of this tool's capabilities.